# python311Packages.loguru: speed-up tests phase