@pbsds 2024-06-13
Please use the full attrpath `python311Packages.loguru` in the commit and PR title to trigger ofborg