@Sigmanificient 2024-06-14
```
error: builder for '/nix/store/x90ihyksfsh0prakm3gx20kksr7fnv40-python3.11-python-utils-3.8.2.drv' failed with exit code 1;
       last 10 log lines:
       >             pass
       >
       > >       assert i == result
       > E       assert 1 == 2
       >
       > _python_utils_tests/test_time.py:84: AssertionError
       > =========================== short test summary info ============================
       > FAILED _python_utils_tests/test_time.py::test_timeout_generator[0.01-0.006-0.5-0.01-count-2] - assert 1 == 2
       > ========================= 1 failed, 33 passed in 4.08s =========================
```
Weird test failling, seems to not be related with loguru logging