@pbsds 2024-06-15
I checked the [test](https://github.com/wolph/python-utils/blob/f6d7e13b9f1d037046a9789c0cf70fcd25b12214/_python_utils_tests/test_time.py#L28), it strikes me as timing sensitivity/flakyness. How did you produce the failure?