@siraben 2024-06-16
Result of `nixpkgs-review pr 317334` run on x86_64-linux [1](https://github.com/Mic92/nixpkgs-review)
<details>
  <summary>2 packages marked as broken and skipped:</summary>
  <ul>
    <li>chatgpt-retrieval-plugin</li>
    <li>chatgpt-retrieval-plugin.dist</li>
  </ul>
</details>
<details>
  <summary>41 packages failed to build:</summary>
  <ul>
    <li>cura</li>
    <li>portmod</li>
    <li>portmod.dist</li>
    <li>python311Packages.angr</li>
    <li>python311Packages.angr.dist</li>
    <li>python311Packages.angrcli</li>
    <li>python311Packages.angrcli.dist</li>
    <li>python311Packages.angrop</li>
    <li>python311Packages.angrop.dist</li>
    <li>python311Packages.awswrangler</li>
    <li>python311Packages.awswrangler.dist</li>
    <li>python311Packages.baselines</li>
    <li>python311Packages.baselines.dist</li>
    <li>python311Packages.gradient</li>
    <li>python311Packages.gradient.dist</li>
    <li>python311Packages.numpy-stl</li>
    <li>python311Packages.numpy-stl.dist</li>
    <li>python311Packages.progressbar2</li>
    <li>python311Packages.progressbar2.dist</li>
    <li>python311Packages.python-utils</li>
    <li>python311Packages.python-utils.dist</li>
    <li>python311Packages.xnatpy</li>
    <li>python311Packages.xnatpy.dist</li>
    <li>python312Packages.awswrangler</li>
    <li>python312Packages.awswrangler.dist</li>
    <li>python312Packages.gradient</li>
    <li>python312Packages.gradient.dist</li>
    <li>python312Packages.numpy-stl</li>
    <li>python312Packages.numpy-stl.dist</li>
    <li>python312Packages.progressbar2</li>
    <li>python312Packages.progressbar2.dist</li>
    <li>python312Packages.python-fx</li>
    <li>python312Packages.python-fx.dist</li>
    <li>python312Packages.python-utils</li>
    <li>python312Packages.python-utils.dist</li>
    <li>python312Packages.xnatpy</li>
    <li>python312Packages.xnatpy.dist</li>
    <li>trueseeing</li>
    <li>trueseeing.dist</li>
    <li>twitch-chat-downloader</li>
    <li>twitch-chat-downloader.dist</li>
  </ul>
</details>
<details>
  <summary>106 packages built:</summary>
  <ul>
    <li>changedetection-io</li>
    <li>changedetection-io.dist</li>
    <li>conjure</li>
    <li>datatrove</li>
    <li>datatrove.dist</li>
    <li>deltachat-cursed</li>
    <li>deltachat-cursed.dist</li>
    <li>flexget</li>
    <li>flexget.dist</li>
    <li>gnome-frog</li>
    <li>home-assistant-component-tests.aussie_broadband</li>
    <li>home-assistant-component-tests.hive</li>
    <li>instawow</li>
    <li>instawow.dist</li>
    <li>ki</li>
    <li>ki.dist</li>
    <li>mokuro</li>
    <li>mokuro.dist</li>
    <li>patray</li>
    <li>patray.dist</li>
    <li>pentestgpt</li>
    <li>pentestgpt.dist</li>
    <li>python311Packages.aria2p</li>
    <li>python311Packages.aria2p.dist</li>
    <li>python311Packages.bubop</li>
    <li>python311Packages.bubop.dist</li>
    <li>python311Packages.ciscoconfparse</li>
    <li>python311Packages.ciscoconfparse.dist</li>
    <li>python311Packages.fastembed</li>
    <li>python311Packages.fastembed.dist</li>
    <li>python311Packages.item-synchronizer</li>
    <li>python311Packages.item-synchronizer.dist</li>
    <li>python311Packages.loguru</li>
    <li>python311Packages.loguru-logging-intercept</li>
    <li>python311Packages.loguru-logging-intercept.dist</li>
    <li>python311Packages.loguru.dist</li>
    <li>python311Packages.manga-ocr</li>
    <li>python311Packages.manga-ocr.dist</li>
    <li>python311Packages.mobi</li>
    <li>python311Packages.mobi.dist</li>
    <li>python311Packages.notify-py</li>
    <li>python311Packages.notify-py.dist</li>
    <li>python311Packages.pinecone-client</li>
    <li>python311Packages.pinecone-client.dist</li>
    <li>python311Packages.pyaussiebb</li>
    <li>python311Packages.pyaussiebb.dist</li>
    <li>python311Packages.pyhiveapi</li>
    <li>python311Packages.pyhiveapi.dist</li>
    <li>python311Packages.pyosohotwaterapi</li>
    <li>python311Packages.pyosohotwaterapi.dist</li>
    <li>python311Packages.pytablewriter</li>
    <li>python311Packages.pytablewriter.dist</li>
    <li>python311Packages.pytest-md-report</li>
    <li>python311Packages.pytest-md-report.dist</li>
    <li>python311Packages.python-fx</li>
    <li>python311Packages.python-fx.dist</li>
    <li>python311Packages.qiling</li>
    <li>python311Packages.qiling.dist</li>
    <li>python311Packages.riscof</li>
    <li>python311Packages.riscof.dist</li>
    <li>python311Packages.riscv-isac</li>
    <li>python311Packages.riscv-isac.dist</li>
    <li>python311Packages.toggl-cli</li>
    <li>python311Packages.toggl-cli.dist</li>
    <li>python312Packages.aria2p</li>
    <li>python312Packages.aria2p.dist</li>
    <li>python312Packages.bubop</li>
    <li>python312Packages.bubop.dist</li>
    <li>python312Packages.ciscoconfparse</li>
    <li>python312Packages.ciscoconfparse.dist</li>
    <li>python312Packages.fastembed</li>
    <li>python312Packages.fastembed.dist</li>
    <li>python312Packages.item-synchronizer</li>
    <li>python312Packages.item-synchronizer.dist</li>
    <li>python312Packages.loguru</li>
    <li>python312Packages.loguru-logging-intercept</li>
    <li>python312Packages.loguru-logging-intercept.dist</li>
    <li>python312Packages.loguru.dist</li>
    <li>python312Packages.mobi</li>
    <li>python312Packages.mobi.dist</li>
    <li>python312Packages.notify-py</li>
    <li>python312Packages.notify-py.dist</li>
    <li>python312Packages.pinecone-client</li>
    <li>python312Packages.pinecone-client.dist</li>
    <li>python312Packages.pyaussiebb</li>
    <li>python312Packages.pyaussiebb.dist</li>
    <li>python312Packages.pyhiveapi</li>
    <li>python312Packages.pyhiveapi.dist</li>
    <li>python312Packages.pyosohotwaterapi</li>
    <li>python312Packages.pyosohotwaterapi.dist</li>
    <li>python312Packages.pytablewriter</li>
    <li>python312Packages.pytablewriter.dist</li>
    <li>python312Packages.pytest-md-report</li>
    <li>python312Packages.pytest-md-report.dist</li>
    <li>python312Packages.riscof</li>
    <li>python312Packages.riscof.dist</li>
    <li>python312Packages.riscv-isac</li>
    <li>python312Packages.riscv-isac.dist</li>
    <li>python312Packages.toggl-cli</li>
    <li>python312Packages.toggl-cli.dist</li>
    <li>resonance</li>
    <li>syncall</li>
    <li>syncall.dist</li>
    <li>varia</li>
    <li>wapiti</li>
    <li>wapiti.dist</li>
  </ul>
</details>